+++ src/main/java/com/klikli_dev/theurgy/integration/modonomicon/page/distillation/BookDistillationRecipePage.java
@@ -24,11 +24,11 @@ public class BookDistillationRecipePage extends BookProcessingRecipePage<Distill
         super(RecipeTypeRegistry.DISTILLATION.get(), title1, recipeId1, title2, recipeId2, text, anchor, condition);
     }
 
-    public static BookDistillationRecipePage fromJson(JsonObject json, HolderLookup.Provider provider) {
+    public static BookDistillationRecipePage fromJson(ResourceLocation entryId, JsonObject json, HolderLookup.Provider provider) {
         var common = BookRecipePage.commonFromJson(json, provider);
         var anchor = GsonHelper.getAsString(json, "anchor", "");
         var condition = json.has("condition")
-                ? BookCondition.fromJson(json.getAsJsonObject("condition"), provider)
+                ? BookCondition.fromJson(entryId, json.getAsJsonObject("condition"), provider)
                 : new BookNoneCondition();
         return new BookDistillationRecipePage(common.title1(), common.recipeId1(), common.title2(), common.recipeId2(), common.text(), anchor, condition);
     }